ADDED functionality to skip initialisation of Navigator ADDED functionality to use own ViewChangeManager
