+++ HybridMenu-addon/src/main/java/kaesdingeling/hybridmenu/HybridMenu.java
@@ -1,8 +1,5 @@
 package kaesdingeling.hybridmenu;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.vaadin.navigator.Navigator;
 import com.vaadin.navigator.ViewChangeListener;
 import com.vaadin.shared.ui.ContentMode;
@@ -12,7 +9,6 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.Layout;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
-
 import kaesdingeling.hybridmenu.components.NotificationCenter;
 import kaesdingeling.hybridmenu.data.DesignItem;
 import kaesdingeling.hybridmenu.data.MenuConfig;
@@ -23,14 +19,18 @@ import kaesdingeling.hybridmenu.data.enums.EMenuDesign;
 import kaesdingeling.hybridmenu.data.leftmenu.MenuButton;
 import kaesdingeling.hybridmenu.data.leftmenu.MenuSubMenu;
 import kaesdingeling.hybridmenu.page.DefaultPage;
+import kaesdingeling.hybridmenu.utils.DefaultViewChangeManager;
 import kaesdingeling.hybridmenu.utils.ViewChangeManager;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class HybridMenu extends VerticalLayout {
 	private static final long serialVersionUID = -4055770717384786366L;
 	
 	public static final String CLASS_NAME = "hybridMenu";
 
-	private ViewChangeManager viewChangeManager = new ViewChangeManager();
+	private ViewChangeManager viewChangeManager = new DefaultViewChangeManager();
 	private MenuConfig config = null;
 	private boolean buildRunning = false;
 	private boolean initNavigator = true;
@@ -296,5 +296,9 @@ public class HybridMenu extends VerticalLayout {
 
 	public void setMenuComponents(EMenuComponents menuComponents) {
 		this.menuComponents = menuComponents;
+	}
+
+	public void setViewChangeManager(ViewChangeManager viewChangeManager) {
+		viewChangeManager = viewChangeManager;
 	}
 }
